Move ReplacedBy to org.gradle.api.model
That seems to be a better package than the top-level org.gradle.api.
